Remove server only code from common/plugin/aws
The agent path construction helpers are only used by the server
plugin now. This change moves that code to keep the common module
as small as possible.
